♻️[Refactor] 코드 구조 및 이름 변경
