+++ src/components/chat-bubble-list/chat-bubble/ChatBubble.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import ChatBubble from './ChatBubble';
 
 const meta = {
-  title: 'Components/ChatBubble/Variants',
+  title: 'Components/ChatList/ChatBubble',
   component: ChatBubble,
   parameters: {
     layout: 'centered',
