+++ src/components/chat-bubble-list/chat-bubble/ChatBubble.tsx
