+++ src/components/chat-bubble-list/chat-bubble/components/OpponentBubble.tsx
