+++ src/components/chat-bubble-list/chat-bubble/components/SystemBubble.tsx
